fix(YouTube - Hide ads): Empty space is left when ads are hidden on tablets
