+++ patches/src/main/kotlin/app/morphe/patches/youtube/layout/searchfilter/Fingerprints.kt
@@ -21,21 +21,6 @@ internal object SearchFilterDialogFingerprint : Fingerprint(
     )
 )
 
-internal object SearchRequestBuildParametersFingerprint : Fingerprint(
-    accessFlags = listOf(AccessFlags.PUBLIC, AccessFlags.FINAL),
-    returnType = "Ljava/lang/String;",
-    parameters = listOf(),
-    filters = listOf(
-        string("searchFormData"),
-        methodCall(
-            opcode = Opcode.INVOKE_VIRTUAL,
-            name = "toByteArray",
-            location = MatchAfterImmediately()
-        ),
-        opcode(Opcode.MOVE_RESULT_OBJECT, location = MatchAfterImmediately()),
-    )
-)
-
 internal object SearchResponseParserFingerprint : Fingerprint(
     accessFlags = listOf(AccessFlags.PROTECTED, AccessFlags.FINAL),
     returnType = "V",
