+++ patches/src/main/kotlin/app/morphe/patches/youtube/misc/contexthook/ClientContextHookPatch.kt
@@ -6,6 +6,7 @@ import app.morphe.patcher.extensions.InstructionExtensions.getInstruction
 import app.morphe.patcher.patch.bytecodePatch
 import app.morphe.patcher.util.proxy.mutableTypes.MutableMethod.Companion.toMutable
 import app.morphe.patches.youtube.misc.extension.sharedExtensionPatch
+import app.morphe.patches.youtube.shared.SearchRequestBuildParametersFingerprint
 import app.morphe.util.addInstructionsAtControlFlowLabel
 import app.morphe.util.findInstructionIndicesReversedOrThrow
 import com.android.tools.smali.dexlib2.AccessFlags
@@ -34,6 +35,7 @@ enum class Endpoint(
         ReelItemWatchEndpointConstructorFingerprint,
         ReelWatchSequenceEndpointConstructorFingerprint,
     ),
+    SEARCH(SearchRequestBuildParametersFingerprint),
     TRANSCRIPT(TranscriptEndpointConstructorFingerprint);
 }
